refactor(rules): split scanner.rs tests into per-area files
Same diagnosis as the execution follow-up: 310 LOC code / 1053 LOC
tests in one file, with the tests already organised into four
self-contained sections separated by `// ── ...` headers. Each
section now lives in its own file:

- scanner/mod.rs                      — code only, declares
                                        `#[cfg(test)] mod tests;`
- scanner/tests/mod.rs                — shared fixtures (make_pack,
                                        test_gates, host_pair,
                                        default_rules) + scanner
                                        integration tests (the bulk —
                                        scan + match_entries happy
                                        path, ignore patterns,
                                        priority ordering, skip
                                        handler, etc.)
- scanner/tests/gates.rs              — Filename-gate tests
                                        (`._<label>` suffix)
- scanner/tests/dir_gates.rs          — C2 directory-segment gates
                                        (`_<label>/`)
- scanner/tests/mapping_gates.rs      — C4 [mappings.gates] glob-based
                                        gates

Helpers are `pub(super)`; per-area files import them via
`use super::{...}`. No production change; same 1140 tests still pass.
